+++ CHANGELOG.md
@@ -42,6 +42,8 @@
     - Added command "Toggle engine pause"
     - Added command "Toggle lock modules"
     - Added LEDs for signaling a triggered hotkey
+- Module [TRANSIT](./docs/Transit.md)
+    - Added trigger-option "pseudo-random"
 - Module [X4](./docs/X4.md)
     - Fixed advancing to the lower button after the upper button has been mapped
     - Fixed wrong tooltip of lower mapping button